apps/strahovka/requirements.txt
alembic==1.4.2
amqp==2.5.2
asgiref==3.2.3
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.8.2
billiard==3.6.3.0
bleach==3.1.1
bottle==0.12.18
bs4==0.0.1
cachetools==4.0.0
celery==4.4.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cycler==0.10.0
DateTime==4.3
debtcollector==2.0.1
decorator==4.4.1
defusedxml==0.6.0
dicttoxml==1.7.4
Django==3.0.2
dnspython==1.16.0
docutils==0.16
dogpile.cache==0.9.0
entrypoints==0.3
eventlet==0.25.1
extras==1.0.0
fake-useragent==0.1.11
fasteners==0.15
fixtures==3.0.0
futurist==2.1.1
get==2019.4.13
gevent==1.4.0
gluon==1.1.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
greenlet==0.4.15
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.17.0
idna==2.8
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.1
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
jupyterlab-server==1.0.6
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keystoneauth1==4.0.0
keystonemiddleware==9.0.0
kiwisolver==1.1.0
kombu==4.6.8
linecache2==1.0.0
lockfile==0.12.2
logutils==0.3.5
lxml==4.5.0
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
monotonic==1.5
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.4
netaddr==0.7.19
netifaces==0.10.9
notebook==6.0.3
numpy==1.18.1
oauth2client==4.1.3
oauthlib==3.1.0
os-service-types==1.7.0
oslo.cache==2.3.0
oslo.concurrency==4.0.2
oslo.config==8.0.2
oslo.context==3.0.2
oslo.db==8.1.0
oslo.i18n==4.0.1
oslo.log==4.1.1
oslo.messaging==12.1.0
oslo.middleware==4.0.2
oslo.policy==3.0.3
oslo.serialization==3.1.1
oslo.service==2.1.1
oslo.utils==4.1.1
oslo.versionedobjects==2.0.2
pandas==1.0.0
pandocfilters==1.4.2
parso==0.6.1
Paste==3.4.0
PasteDeploy==2.1.0
pbr==5.4.5
pecan==1.3.3
pickleshare==0.7.5
Pillow==7.0.0
pluralize==0.1.5
post==2019.4.13
prometheus-client==0.7.1
prompt-toolkit==3.0.3
psutil==5.7.0
public==2019.4.13
py4web==0.1.20200202.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycadf==3.0.0
pycparser==2.19
pydal==20200404.1
PyDrive==1.3.1
Pygments==2.5.2
PyJWT==1.7.1
pyparsing==2.4.6
pyrsistent==0.15.7
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-etcd==0.4.5
python-keystoneclient==3.22.0
python-memcached==1.59
python-mimeparse==1.6.0
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==19.0.0
query-string==2019.4.13
redis==3.4.1
repoze.lru==0.7
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
rfc3986==1.3.2
Routes==2.4.1
rsa==4.0
scipy==1.4.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.0.19
sqlalchemy-migrate==0.13.0
sqlparse==0.3.0
statsd==3.3.0
stevedore==1.32.0
Tempita==0.5.2
terminado==0.8.3
testpath==0.4.4
testresources==2.0.1
testscenarios==0.5.0
testtools==2.4.0
threadsafevariable==1.0
tornado==6.0.3
traceback2==1.4.0
traitlets==4.3.3
unittest2==1.1.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
vprof==0.38
waitress==1.4.3
watchgod==0.6
wcwidth==0.1.8
webencodings==0.5.1
WebOb==1.8.6
WebTest==2.0.34
wrapt==1.12.1
WSME==0.9.3
xlrd==1.2.0
xmltodict==0.12.0
yappi==1.2.3
yatl==20200329.1
zope.interface==4.7.1
